+++ src/main/java/com/example/GestionEmployes/Repository/EmployeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.example.GestionEmployes.Models.Employe;
 
-public interface EmployeRepository extends CrudRepository<Employe, Integer>{
+public interface EmployeRepository extends CrudRepository<Employe, Long>{
 
 }
